refactor: extract query-types crate from query (Part 1 of #670)
Move mapper/, document/, error.rs, json_convert.rs, and the Doc/DocStatus
data types from the query crate into a new query-types crate (~10K lines).

These modules define the shared type vocabulary (Filter, Select, Mutation,
DocumentMapping, Doc, QueryError) used across parsing, planning, and
execution. Extracting them enables Part 2: extraction of query-parse
without circular dependencies.

The query crate re-exports all types via module aliases so all downstream
imports continue to work unchanged.
