+++ Cargo.toml
@@ -8,6 +8,7 @@ members = [
     "crates/embedded",
     "crates/schema",
     "crates/query",
+    "crates/query-types",
     "crates/document",
     "crates/p2p",
     "crates/crypto",
