+++ crates/query-types/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "query-types"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+license.workspace = true
+repository.workspace = true
+description = "Shared type definitions for the DefraDB query engine"
+
+[dependencies]
+# Serialization
+serde.workspace = true
+serde_json.workspace = true
+
+# Error handling
+thiserror.workspace = true
+
+# Logging
+tracing.workspace = true
+
+# Time
+chrono = { version = "0.4", features = ["serde"] }
+
+# Internal crates
+schema = { path = "../schema" }
+storage = { path = "../storage" }
+document = { path = "../document" }
+defra-core = { path = "../defra-core" }
+
+[dev-dependencies]
+tokio = { workspace = true, features = ["rt-multi-thread", "macros"] }
