+++ crates/query-types/src/mapper/filter/op.rs
@@ -5,7 +5,6 @@ use serde::{Deserialize, Serialize};
 /// Filter operators for condition matching
 /// Uses Go DefraDB naming conventions for compatibility
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash, Serialize, Deserialize)]
-#[non_exhaustive]
 pub enum FilterOp {
     /// Equal (_eq)
     #[serde(rename = "_eq")]
